requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
psycopg2-binary==2.9.6
sqlalchemy==1.4.41
alembic==1.8.1 
mgrs==1.4.5
Pillow==10.4.0